Conflicts:
	src/cosc561/tsp/Solver.java
	src/cosc561/tsp/view/MapWindow.java

+++ src/cosc561/tsp/strategy/BranchAndBoundPermutation.java
@@ -13,7 +13,7 @@ public class BranchAndBoundPermutation extends Strategy {
 	
 	private static final int QUEUE_SIZE = 1000;
 
-	private int rejected;
+	private long rejected;
 
 	private Queue<SparseBranch> permutationsInProgress;
